+++ libraries/app/test/app_test.js
@@ -4,6 +4,9 @@ const request = require('superagent');
 const express = require('express');
 const isUUID = require('is-uuid');
 const testing = require('taskcluster-lib-testing');
+const path = require('path');
+const rootdir = require('app-root-dir');
+const mockFs = require('mock-fs');
 
 suite(testing.suiteName(), function() {
 
@@ -79,6 +82,31 @@ suite(testing.suiteName(), function() {
       assert.equal(res.headers['content-type'], 'text/plain; charset=utf-8');
     });
 
+    test('/__version__', async function() {
+      mockFs({
+        [path.resolve(rootdir.get(), 'version.json')]: JSON.stringify({ version: 'v99.99.99' }),
+      });
+
+      const res = await request.get('http://localhost:1459/__version__');
+      assert(res.ok, 'Got response');
+      assert.equal(res.body.version, 'v99.99.99', 'Got the right version');
+      assert.equal(res.headers['content-type'], 'application/json; charset=utf-8');
+    });
+
+    test('/__heartbeat__', async function() {
+      const res = await request.get('http://localhost:1459/__heartbeat__');
+      assert(res.ok, 'Got response');
+      assert.equal(res.status, 200);
+      assert.equal(res.headers['content-type'], 'application/json; charset=utf-8');
+    });
+
+    test('/__lbheartbeat__', async function() {
+      const res = await request.get('http://localhost:1459/__lbheartbeat__');
+      assert(res.ok, 'Got response');
+      assert.equal(res.status, 200);
+      assert.equal(res.headers['content-type'], 'application/json; charset=utf-8');
+    });
+
     test('/not-found', async function() {
       try {
         await request.get('http://localhost:1459/api/test/v1/notfound');
@@ -94,6 +122,10 @@ suite(testing.suiteName(), function() {
         return;
       }
       throw new Error('expected exception not seen');
+    });
+
+    teardown(function() {
+      mockFs.restore();
     });
 
     suiteTeardown(function() {