+++ libraries/entities/test/row_modify_test.js
@@ -34,6 +34,29 @@ helper.dbSuite(path.basename(__filename), function() {
   const serviceName = 'test-entities';
 
   suite('row modify', function() {
+    test('modify without changing anything should not call the _modify function', async function() {
+      db = await helper.withDb({ schema, serviceName });
+      const id = slugid.v4();
+      const TestTable = configuredTestTable.setup({ tableName: 'test_entities', db, serviceName });
+
+      const item = await TestTable.create({
+        id: id,
+        name: 'my-test-item',
+        count: 1,
+        time: new Date(),
+      });
+
+      const oldModify = db.fns.test_entities_modify;
+      db.fns.test_entities_modify = () => { throw new Error('should not happen'); };
+      try {
+        const item2 = await item.modify(function(entry) {
+          // do nothing
+        });
+        assert.deepEqual(item, item2);
+      } finally {
+        db.fns.test_entities_modify = oldModify;
+      }
+    });
     test('Item.create, Item.modify, Item.load', async function() {
       db = await helper.withDb({ schema, serviceName });
       const id = slugid.v4();