+++ services/web-server/test/graphql/workermanager_test.js
@@ -4,7 +4,7 @@ const testing = require('taskcluster-lib-testing');
 const helper = require('../helper');
 const deleteWorkerPoolMutation = require('../fixtures/deleteWorkerPool.graphql');
 
-helper.secrets.mockSuite(testing.suiteName(), [], function(mock, skipping) {
+helper.secrets.mockSuite(testing.suiteName(), ['db'], function(mock, skipping) {
   helper.withDb(mock, skipping);
   helper.withEntities(mock, skipping);
   helper.withClients(mock, skipping);